Add files via upload
Written by Robert C. Ryniak, July 27, 2016 for SoftWriters. Estimated 4 hours.  Written in C# using Visual Studio 2010.  The meat of the independent functionality as required is in the file OnScreenKeyMapper.cs.  Please note that this work isn't about showing you how well I know a particular programming language, or the latest frameworks, or any of the things which seem to be in constant flux in the world of software engineering.  It is the purpose and goal of this code to demonstrate how I solve the problem in the "requirements" as provided by the stakeholder(s). If you have any questions why I did things one particular way or another, please don't hesitate to ask me, thanks!